feat: entity configurations moved
