+++ src/Infrastructure/Persistence/Configurations/Identity/UserTokenConfiguration.cs
@@ -1,0 +1,20 @@
+﻿using BB84.EntityFrameworkCore.Repositories.SqlServer.Extensions;
+
+using Domain.Entities.Identity;
+
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+using static Infrastructure.Common.InfrastructureConstants;
+
+namespace Infrastructure.Persistence.Configurations.Identity;
+
+/// <summary>
+/// The configuration for the <see cref="UserTokenModel"/> entity.
+/// </summary>
+internal class UserTokenConfiguration : IEntityTypeConfiguration<UserTokenModel>
+{
+	/// <inheritdoc/>
+	public void Configure(EntityTypeBuilder<UserTokenModel> builder) =>
+		builder.ToHistoryTable("UserToken", SqlSchema.Identity, SqlSchema.History);
+}
