--- src/Infrastructure/Persistence/Configurations/IdentityConfiguration.cs
@@ -1,7 +1,0 @@
-﻿namespace Infrastructure.Persistence.Configurations;
-
-/// <summary>
-/// The <see langword="static"/> identity configuration class.
-/// </summary>
-internal static partial class IdentityConfiguration
-{ }
